fix: add new treasury structure to tests
This implements the -mining and -stacking versions of the treasury, replacing any related code with the version it called (in most cases - mining).

For ccd002 testing any one of the treasury contracts could be used, since they all share the same code. For ccd006 it should refer to the -mining variants. For ccd007 to -stacking.

Testing multiple contracts (esp. with Clarinet requirements), and testing across cities are both good candidates for fuzzing!