fix: update naming in activation details
The original language used activated to represent the block that the final user registered, followed by the target block which would be activation + delay.

This was confusing in the tests and when comparing to the original MIA/NYC data passed in by CCIP-013.

Instead we can use succeeded for the block that the final user registered, and activated will represent the succeeded + delay, so that activated means activated in the context of the protocol.