fix: add coinbase details for bonus and epoch
The bonus period was being tracked in a separate map but the epoch length was not, which is required for new city activations and determining future block heights based on the emissions schedule.

This combines the two and returns it via the grouped get-coinbase-info, and also adds a test to make sure neither value can be 0.