fix: add trait support and disabled trait functions to mining contract
Using the citycoins-core-v2 trait allows ccd006 to be considerd a "core contract" by the existing auth-v2 contracts, which allows for some backwards-compatibility with changes that need to be made in the old protocol, mainly in the token contracts.

BREAKING CHANGE: the existing claim-mining-reward function is in conflict with the original trait specification, and is now renamed "claim-mining-block". This is updated in both contracts as well as all related tests.