fix: add adapter for v2-auth functions
This extension contract should be set as an approver in the auth-v2 contracts for MIA/NYC, then can be used by the DAO to initiate or execute the auth jobs in the context of the DAO.

The primary use case for this is to both shut down the original protocol and start the new one in the same transaction/proposal.

This also adds some functions to ccd006 for compatibility with the original core contract specification.