fix: resolve test errors mostly around block heights
There were a few spots hardcoded to a constant start value, but the clarinet test environment doesn't guarantee this will be the same block every time.

By setting this to the relative value based on the processed block, the results come out the same every time, even if the underlying block numbers change ever so slightly.