fix: update print ordering to be consistent
This makes it so that every print function starts with "event: name" then all items that follow are in ABC order.

The name is the main internal designation for the type of event, however all items within the print event are returned in alphabetical order from the API and external sources.