fix: add outline for ccip-013 tests
The test here will succeed for the first two transactions against the base-dao, but fails when the proposal is passed since the MIA/NYC legacy protocol is not activated in the test suite.

There should be come code we can port over from the citycoins/contracts repo that covered the v1->v2 migration testing to finish out the requirements here, but the test code is structured differently so the models may need to be rewritten.