fix: update ccd002 stacking functions
This changes the naming to be clearer as to what each function calls in PoX. The variables have been simplified as well as the pool operator is responsible for the pox address if not specified.

This also adds tests for the success case of each using a ccd002 test proposal.