fix (wip): weird errors on edge cases but overall structure in place
Something is happening with inheritance when checking tx-sender as part of both functions in ccd011 and ccd007.

The tests were trying to use a proposal to send the stacking reward but this turned out to be unneeded complexity as the ccd007 test 001 covers setting the pool admin to the deployer address.