fix: add traits for ccd002/006/007
This implements a basic trait for each part of the protocol, which should be helpful as the protocol evolves beyond the first version.

This also experimented with removing get-block-winner functions and related map, but cost gains were marginal and given others voiced the usefulness in front-ends, it will remain for now.