fix: bring over changes from 1ee6ce9 in feat/mainnet-deployment
There was an addition of a boolean flag for allowing future cycle claims if the contract is disabled, guarded by is-dao-or-extension and set to false by default.

This incorporates the same change to ensure all tests pass and syntax is valid.

In addition, this removes the check for a city treasury in ccd007 set-stacking-reward since the transfer function and destination is now handled by ccd011, and removes an extra print statement used in testing.