fix: create clone of ccip-014 as v2
The original proposal failed with err u7006 due to cycle 59 not being complete. This would not allow us to stack for cycle 60, so the related code will be removed and saved for a future proposal.

The voting functionality will be removed as well and instead of checking internally if it can be executed, it will call the original ccip-014.