fix: update testing strategy
If we change epoch from 2.5 to 2.4 there are no changes to Clarity and older test structure can still be used. This will allow codecov to capture the changes for the new contracts and unifies everything back under a single Clarinet.toml file.

The GH action is updated to run clarinet check with both legacy and latest versions, as well as run the legacy tests with a clear note by each job name.