fix: update test to iterate and check contract, user, and ratios
This expands the final test to check and log the data before and after key events in the redemption process, as well as check the redemption ratio both at the contract level and with the redemption for each user. Console logging is left on for now to help others in review.