fix: support edge case of claim being greater than contract balance
This was identified in peer review, since mining claims are still available there is a slight chance that the amount claimed by the final claimant may be greater than the amount of STX in the contract. As a result, the user would be required to reduce their NYC balance to be able to claim the remaining STX in the contract, which is now mitigated by the added IF statement.