fix: adjust block height to be in future for tests
All tests are expecting to use early stacking cycles in order to vote, but the contract sets the voting window too soon when it just goes off the block height it's deployed. For testing we extend the set blocks by u12600 which represents 6 * 2100 block cycle length. This should be removed for mainnet deployment and comments above include the correct code for deployment.